openvpn: ensure that we restore previous DNS settings after shutting down a client in case the updown.sh script wasn't run.
Certain OpeNVPN client configurations (with explicit-exit-notify enabled)
will cause the openvpn server to skip executing the --down script,
leading to customized DNS settings not to be reverted.  This will resolve
the case where the end-user is the one stopping the client, but it might
not work properly for cases where the tunnel has unexpectedly been
terminated, or terminated by the server.